Move all types in one types.ts file
