+++ packages/shopify-app-session-storage/src/abstract-migration-engine.ts
@@ -4,8 +4,8 @@ import {
   SessionStorageMigratorOptions,
   defaultSessionStorageMigratorOptions,
   InvalidMigrationConfigurationError,
-} from './session-storage-migration';
-import {DBConnection} from './db-connection';
+  DBConnection,
+} from './types';
 
 export abstract class AbstractMigrationEngine<
   ConnectionType extends DBConnection,
