+++ packages/shopify-app-session-storage/src/index.ts
@@ -1,5 +1,3 @@
-export * from './session-storage';
-export * from './session-storage-migration';
+export * from './types';
 export * from './rdbms-session-storage-migrator';
 export * from './abstract-migration-engine';
-export * from './db-connection';
